requirements.txt
feedparser>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.3
openai>=1.12.0
apscheduler==3.10.4
python-dotenv==1.0.1
google-search-results>=2.4.2  # SerpAPI client
python-slugify>=5.0.2
lxml>=4.9.0
Pillow>=9.0.0  # For image processing
tweepy==4.14.0  # For Twitter API integration
mailchimp3==3.0.19
python-dateutil==2.8.2
elevenlabs>=0.2.27
ghost-admin-api>=0.1.0
linkedin-api==2.0.0a5  # NOTE: This library may be unmaintained; consider finding an alternative
praw>=7.7.1
ttkbootstrap==1.10.1
eel>=0.16.0 